feat: add news listing page
