+++ pages/about/news/[slug].vue
